SmartTelegram/pyproject.toml
[project]
name = "smarttelegram"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "aiogram>=3.27.0",
    "python-dotenv>=1.2.2",
]
